requirements.txt
python-dotenv==1.1.1
cryptography==43.0.1
nicegui==3.6.1